requirements.txt
agno
mcp
groq
python-dotenv
